- fixed compilation warning with GCC and Clang
src/hwrenderer/dynlights/hw_aabbtree.cpp:143:24: warning: comparison of integers of different signs: 'const int' and 'unsigned int' [-Wsign-compare]

+++ src/hwrenderer/dynlights/hw_aabbtree.cpp
@@ -140,7 +140,7 @@ TArray<int> LevelAABBTree::FindNodePath(unsigned int line, unsigned int node)
 		if (path.Size())
 			path.Push(node);
 	}
-	else if (n.line_index == line)
+	else if (n.line_index == (int)line)
 	{
 		path.Push(node);
 	}
